simplified and moved deploy-commands to utility.js. deploying commands on startup
